Flask_Library/requirements.txt
Flask-smorest
Flask
Python.dotenv
marshmallow
Flask-SQLAlchemy